edgar_agent/requirements.txt
google-adk
finnhub-python
requests
beautifulsoup4
google-cloud-secret-manager
google-auth
sec-api